% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionMaps.R
\name{affinity}
\alias{affinity}
\title{Gaussian affinity matrix constructor}
\usage{
affinity(R, k = 7, sigma = NULL, log = FALSE, normalize = FALSE)
}
\arguments{
\item{R}{symmetric matrix; distance or any similarity matrix.}

\item{k}{integer; number of neighbours in adaptive-scaling. (ignored if sigma is specified)}

\item{sigma}{numeric; fixed sigma for isotropic diffusion}

\item{log}{logical; if TRUE, log-transformed affinity matrix is returned.}

\item{normalize}{logical; scale s.t. row summing to 1 (i.e. returns transition matrix)}
}
\description{
equivalent to diffusion maps (Coifman, 2015) with adaptive scaling according to local density.
}
\references{
"Self-tuning spectral clustering"
}
