% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphConstr.R
\name{diffusionGraph}
\alias{diffusionGraph}
\title{Construction of diffusion graph}
\usage{
diffusionGraph(X, roots, k = 11, npc = min(100, dim(X) - 1),
  ndc = 40, s = 1, j = 7, lambda = 1e-04, sigma = NULL)
}
\arguments{
\item{X}{data matrix (columns: n samples, rows: p variables)}

\item{roots}{indices of columns of X or logical vector; start points of diffusion (i.e., the initial density is 1/n at specified n points).}

\item{k}{integer; number of nearest neighbours to construct initial k-NN graph}

\item{npc}{integer; number of principal components in calculating euclid distance.}

\item{ndc}{integer; number of diffusion components in calculating diffusion distance.}

\item{s}{numeric; time parameter of diffusion maps.}

\item{j}{integer; use j-th nearest neighbouor in local scaling of sigma}

\item{lambda}{numeric; ridge penalty in pulling back divergence}

\item{sigma}{numeric; fixed sigma for isotropic diffusion}
}
\description{
Construct graph from data matrix based on diffusion maps (Coifman, 2015) with adaptive scaling according to local density.
}
\references{
"Diffusion maps"

"Self-tuning spectral clustering"
}
