% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionMaps.R
\name{diffusionMaps}
\alias{diffusionMaps}
\title{Diffusion maps}
\usage{
diffusionMaps(R, k = 7, sigma = NULL)
}
\arguments{
\item{R}{distance or any similarity matrix.}

\item{k}{number of neighbours in adaptive-scaling. (ignored if sigma is specified)}

\item{sigma}{fixed sigma for isotopic diffusion}
}
\description{
equivalent to diffusion maps (Coifman, 2015) with adaptive scaling according to local density.
}
\references{
"Self-tuning spectral clustering"
}
