% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgeDecomp.R
\name{grad}
\alias{grad}
\title{Calculate gradient}
\usage{
grad(g, bias = 0)
}
\arguments{
\item{g}{igraph object}

\item{bias}{bias term in ridge regression}
}
\description{
graph -> numeric vector
Extract gradient flow by orthogonal projection
}
