% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphConstr.R
\name{multimaxflow}
\alias{multimaxflow}
\title{Multi-source multi-sink maximum flow}
\usage{
multimaxflow(g, source, sink)
}
\arguments{
\item{g}{igraph object}

\item{source}{sourse(s)}

\item{sink}{sink(s)}
}
\value{
igraph object with `E(g)$flow` and `V(g)$pass` attributes.
}
\description{
Simple extension of `max_flow` for multi-source/sink.
}
