% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgeDecomp.R
\name{potential}
\alias{potential}
\title{Potential of nodes}
\usage{
potential(g, bias = 0)
}
\arguments{
\item{g}{igraph object}

\item{bias}{bias term in ridge regression}
}
\description{
some description
}
