% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphConstr.R
\name{diffusionGraph}
\alias{diffusionGraph}
\title{Construction of diffusion graph}
\usage{
diffusionGraph(X, roots, j = 7, k = 11, s = 1, ndc = 10,
  npc = min(1000, dim(X)), lambda = 1e-04, alpha = 0.5,
  sigma = NULL)
}
\arguments{
\item{X}{data matrix (column: samples, rows: variables)}

\item{roots}{indices of columns of X or logical vector; start points of diffusion (i.e., the initial density is 1/n at specified n points).}

\item{j}{integer; use j-th nearest neighbouor in local scaling of sigma}

\item{k}{integer; number of nearest neighbours to construct initial k-NN graph}

\item{s}{numeric; time parameter of diffusion maps.}

\item{ndc}{integer; number of diffusion components in calculating diffusion distance.}

\item{npc}{integer; number of principal components in calculating euclid distance.}

\item{lambda}{numeric; LASSO penalty of elasticnet}

\item{alpha}{numeric; mixing perameter of ridge:0 and LASSO:1 (defaut: 0.5)}

\item{sigma}{numeric; fixed sigma for isotropic diffusion}
}
\description{
Construct graph from data matrix based on diffusion maps (Coifman, 2015) with adaptive scaling according to local density.
}
\references{
"Diffusion maps"

"Self-tuning spectral clustering"
}
