% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgeDecomp.R
\name{potential}
\alias{potential}
\title{Calculate potential}
\usage{
potential(g, tol = 1e-07)
}
\arguments{
\item{g}{igraph object}

\item{tol}{tolerance for solve.QR}
}
\description{
graph -> numeric vector
}
